Enum insta::internals::Content

pub enum Content {

    Bool(bool),
    U8(u8),
    U16(u16),
    U32(u32),
    U64(u64),
    U128(u128),
    I8(i8),
    I16(i16),
    I32(i32),
    I64(i64),
    I128(i128),
    F32(f32),
    F64(f64),
    Char(char),
    String(String),
    Bytes(Vec<u8>),
    Seq(Vec<Content>),
    Map(Vec<(Content, Content)>),
    // some variants omitted

}

Represents variable typed content.

This is used for the serialization system to represent values before the actual snapshots are written and is also exposed to dynamic redaction functions.

Some enum variants are intentionally not exposed to user code. It’s generally recommended to construct content objects by using the From trait and by using the accessor methods to assert on it.

While matching on the content is possible in theory it is recommended against. The reason for this is that the content enum holds variants that can “wrap” values where it’s not expected. For instance if a field holds an Option<String> you cannot use pattern matching to extract the string as it will be contained in an internal Some variant that is not exposed. On the other hand the as_str method will automatically resolve such internal wrappers.

If you do need to pattern match you should use the resolve_inner method to resolve such internal wrappers.

Variants

Bool(bool)

U8(u8)

U16(u16)

U32(u32)

U64(u64)

U128(u128)

I8(i8)

I16(i16)

I32(i32)

I64(i64)

I128(i128)

F32(f32)

F64(f64)

Char(char)

String(String)

Bytes(Vec<u8>)

Seq(Vec<Content>)

Map(Vec<(Content, Content)>)

Implementations

impl Content

pub fn resolve_inner(&self) -> &Content

This resolves the innermost content in a chain of wrapped content.

For instance if you encounter an Option<Option<String>> field the content will be wrapped twice in an internal option wrapper. If you need to pattern match you will need in some situations to first resolve the inner value before such matching can take place as there is no exposed way to match on these wrappers.

This method does not need to be called for the as_ methods which resolve automatically.

pub fn resolve_inner_mut(&mut self) -> &mut Content

Mutable version of resolve_inner.

pub fn as_str(&self) -> Option<&str>

Returns the value as string

pub fn as_bytes(&self) -> Option<&[u8]>

Returns the value as bytes

pub fn as_slice(&self) -> Option<&[Content]>

Returns the value as slice of content values.

pub fn is_nil(&self) -> bool

Returns true if the value is nil.

pub fn as_bool(&self) -> Option<bool>

Returns the value as bool

pub fn as_u64(&self) -> Option<u64>

Returns the value as u64

pub fn as_u128(&self) -> Option<u128>

Returns the value as u128

pub fn as_i64(&self) -> Option<i64>

Returns the value as i64

pub fn as_i128(&self) -> Option<i128>

Returns the value as i128

pub fn as_f64(&self) -> Option<f64>

Returns the value as f64

pub fn walk<F: FnMut(&mut Content) -> bool>(&mut self, visit: &mut F)

Recursively walks the content structure mutably.

The callback is invoked for every content in the tree.