insta
Module runtime

Structs

AutoName

Special marker to use an automatic name.

BinarySnapshotValue

InlineValue

SnapshotAssertionContext 🔒

The context around a snapshot, such as the reference value, location, etc. (but not including the generated value). Responsible for saving the snapshot.

Enums

SnapshotValue

Constants

RECORDED_DUPLICATES 🔒

_DOCTEST1 🔒

Test snapshots in doctests.

Statics

INLINE_DUPLICATES 🔒

TEST_NAME_CLASH_DETECTION 🔒

TEST_NAME_COUNTERS 🔒

Functions

add_suffix_to_snapshot_name 🔒

If there is a suffix on the settings, append it to the snapshot name.

allow_duplicates 🔒

Do we allow recording of duplicates?

assert_snapshot

This function is invoked from the macros to run the main assertion logic.

detect_snapshot_name 🔒

get_snapshot_filename 🔒

is_doctest 🔒

prevent_inline_duplicate 🔒

record_snapshot_duplicate 🔒

with_allow_duplicates

Helper function to support perfect duplicate detection.

Type Aliases

SnapshotName 🔒

The name of a snapshot, from which the path is derived.